Adding intrinsic checking for signature
Most sketches can now simply initialize the library by calling the Begin
function which tests the EEPROM for a signature (a checksum generated
from the sizes of the items) and retrieves the values if it matches; if
the checksum doesn't match, the Begin function will write the defaults
and the checksum by default.

Also moved some code to the .cpp file, and did some clean-up.